conditionally update rubies

+++ hooks/post-up/latest-rubies.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
-ruby-install -L > /dev/null
+if [ "$DOTFILE_UPDATE" = "force" ]; then
+  echo "[updating rubies]"
+  ruby-install -L > /dev/null
+else
+  echo "[not updating rubies]"
+fi
